+++ emmet-core/emmet/core/vasp/material.py
@@ -6,13 +6,12 @@ from pymatgen.analysis.structure_analyzer import SpacegroupAnalyzer
 from pymatgen.analysis.structure_matcher import StructureMatcher
 from pymatgen.entries.computed_entries import ComputedStructureEntry
 
-from emmet.core.settings import EmmetSettings
 from emmet.core.material import MaterialsDoc as CoreMaterialsDoc
 from emmet.core.material import PropertyOrigin
+from emmet.core.settings import EmmetSettings
 from emmet.core.structure import StructureMetadata
 from emmet.core.vasp.calc_types import CalcType, RunType, TaskType
 from emmet.core.vasp.task_valid import TaskDocument
-
 
 SETTINGS = EmmetSettings()
 
@@ -44,7 +43,9 @@ class MaterialsDoc(CoreMaterialsDoc, StructureMetadata):
     def from_tasks(
         cls,
         task_group: List[TaskDocument],
-        quality_scores: Dict[str, int] = SETTINGS.VASP_QUALITY_SCORES,
+        structure_quality_scores: Dict[
+            str, int
+        ] = SETTINGS.VASP_STRUCTURE_QUALITY_SCORES,
         use_statics: bool = SETTINGS.VASP_USE_STATICS,
     ) -> "MaterialsDoc":
         """
@@ -52,7 +53,7 @@ class MaterialsDoc(CoreMaterialsDoc, StructureMetadata):
 
         Args:
             task_group: List of task document
-            quality_scores: quality scores for various calculation types
+            structure_quality_scores: quality scores for various calculation types
             use_statics: Use statics to define a material
         """
         if len(task_group) == 0:
@@ -87,11 +88,11 @@ class MaterialsDoc(CoreMaterialsDoc, StructureMetadata):
         material_id = min(possible_mat_ids)
 
         # Always prefer a static over a structure opt
-        task_quality_scores = {"Structure Optimization": 1, "Static": 2}
+        structure_task_quality_scores = {"Structure Optimization": 1, "Static": 2}
 
         def _structure_eval(task: TaskDocument):
             """
-            Helper function to order structures optimziation and statics calcs by
+            Helper function to order structures optimization and statics calcs by
             - Functional Type
             - Spin polarization
             - Special Tags
@@ -106,8 +107,8 @@ class MaterialsDoc(CoreMaterialsDoc, StructureMetadata):
 
             return (
                 -1 * int(task.is_valid),
-                -1 * quality_scores.get(task_run_type.value, 0),
-                -1 * task_quality_scores.get(task.task_type.value, 0),
+                -1 * structure_quality_scores.get(task_run_type.value, 0),
+                -1 * structure_task_quality_scores.get(task.task_type.value, 0),
                 -1 * special_tags,
                 task.output.energy_per_atom,
             )
@@ -139,6 +140,30 @@ class MaterialsDoc(CoreMaterialsDoc, StructureMetadata):
 
         # Entries
         # **current materials docs must contain at last one GGA or GGA+U entry
+
+        # Always prefer a static over a structure opt
+        entry_task_quality_scores = {"Structure Optimization": 1, "Static": 2}
+
+        def _entry_eval(task: TaskDocument):
+            """
+            Helper function to order entries and statics calcs by
+            - Spin polarization
+            - Special Tags
+            - Energy
+            """
+
+            _SPECIAL_TAGS = ["LASPH", "ISPIN"]
+            special_tags = sum(
+                task.input.parameters.get(tag, False) for tag in _SPECIAL_TAGS
+            )
+
+            return (
+                -1 * int(task.is_valid),
+                -1 * entry_task_quality_scores.get(task.task_type.value, 0),
+                -1 * special_tags,
+                task.output.energy_per_atom,
+            )
+
         entries = {}
         all_run_types = set(run_types.values())
 
@@ -150,7 +175,7 @@ class MaterialsDoc(CoreMaterialsDoc, StructureMetadata):
         for rt in all_run_types:
             relevant_calcs = sorted(
                 [doc for doc in structure_calcs if doc.run_type == rt and doc.is_valid],
-                key=_structure_eval,
+                key=_entry_eval,
             )
 
             if len(relevant_calcs) > 0:
@@ -178,7 +203,7 @@ class MaterialsDoc(CoreMaterialsDoc, StructureMetadata):
 
     @classmethod
     def construct_deprecated_material(
-        cls, task_group: List[TaskDocument],
+        cls, task_group: List[TaskDocument]
     ) -> "MaterialsDoc":
         """
         Converts a group of tasks into a deprecated material