+++ docs/CHANGELOG.md
@@ -1,5 +1,166 @@
 # Changelog
 
+## [v0.30.5](https://github.com/materialsproject/emmet/tree/v0.30.5) (2022-07-13)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.30.4...v0.30.5)
+
+**Merged pull requests:**
+
+- Speed up `TaskValidation` [\#458](https://github.com/materialsproject/emmet/pull/458) ([munrojm](https://github.com/munrojm))
+
+## [v0.30.4](https://github.com/materialsproject/emmet/tree/v0.30.4) (2022-07-13)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.30.3...v0.30.4)
+
+**Merged pull requests:**
+
+- Validation bugfix [\#457](https://github.com/materialsproject/emmet/pull/457) ([munrojm](https://github.com/munrojm))
+
+## [v0.30.3](https://github.com/materialsproject/emmet/tree/v0.30.3) (2022-07-13)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.30.2...v0.30.3)
+
+**Merged pull requests:**
+
+- Build POTCAR hash cache in validation builder [\#456](https://github.com/materialsproject/emmet/pull/456) ([munrojm](https://github.com/munrojm))
+
+## [v0.30.2](https://github.com/materialsproject/emmet/tree/v0.30.2) (2022-07-12)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.30.1...v0.30.2)
+
+**Merged pull requests:**
+
+- Change structure quality scores for materials builder [\#455](https://github.com/materialsproject/emmet/pull/455) ([munrojm](https://github.com/munrojm))
+
+## [v0.30.1](https://github.com/materialsproject/emmet/tree/v0.30.1) (2022-07-12)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.30.0...v0.30.1)
+
+**Merged pull requests:**
+
+- Fix handling deformation tasks in materials builder [\#453](https://github.com/materialsproject/emmet/pull/453) ([munrojm](https://github.com/munrojm))
+
+## [v0.30.0](https://github.com/materialsproject/emmet/tree/v0.30.0) (2022-07-12)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.29.5...v0.30.0)
+
+**Merged pull requests:**
+
+- Migrationgraphdoc [\#437](https://github.com/materialsproject/emmet/pull/437) ([hmlli](https://github.com/hmlli))
+
+## [v0.29.5](https://github.com/materialsproject/emmet/tree/v0.29.5) (2022-07-12)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.29.4...v0.29.5)
+
+**Merged pull requests:**
+
+- Validation fixes [\#454](https://github.com/materialsproject/emmet/pull/454) ([munrojm](https://github.com/munrojm))
+
+## [v0.29.4](https://github.com/materialsproject/emmet/tree/v0.29.4) (2022-07-08)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.29.3...v0.29.4)
+
+**Merged pull requests:**
+
+- Properly set potcar dir in validation [\#450](https://github.com/materialsproject/emmet/pull/450) ([munrojm](https://github.com/munrojm))
+
+## [v0.29.3](https://github.com/materialsproject/emmet/tree/v0.29.3) (2022-07-07)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.29.2...v0.29.3)
+
+**Merged pull requests:**
+
+- Fix builder settings [\#449](https://github.com/materialsproject/emmet/pull/449) ([munrojm](https://github.com/munrojm))
+
+## [v0.29.2](https://github.com/materialsproject/emmet/tree/v0.29.2) (2022-07-07)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.29.1...v0.29.2)
+
+**Merged pull requests:**
+
+- Add mobility utility functions [\#447](https://github.com/materialsproject/emmet/pull/447) ([acrutt](https://github.com/acrutt))
+
+## [v0.29.1](https://github.com/materialsproject/emmet/tree/v0.29.1) (2022-07-07)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.29.0...v0.29.1)
+
+**Merged pull requests:**
+
+- Materials builder scan changes [\#448](https://github.com/materialsproject/emmet/pull/448) ([munrojm](https://github.com/munrojm))
+
+## [v0.29.0](https://github.com/materialsproject/emmet/tree/v0.29.0) (2022-07-06)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.28.4...v0.29.0)
+
+**Merged pull requests:**
+
+- Add optic absorption builder  [\#425](https://github.com/materialsproject/emmet/pull/425) ([yang-ruoxi](https://github.com/yang-ruoxi))
+
+## [v0.28.4](https://github.com/materialsproject/emmet/tree/v0.28.4) (2022-06-27)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.28.3...v0.28.4)
+
+**Merged pull requests:**
+
+- Ensure electrode groups only pull non-deprecated materials [\#443](https://github.com/materialsproject/emmet/pull/443) ([munrojm](https://github.com/munrojm))
+
+## [v0.28.3](https://github.com/materialsproject/emmet/tree/v0.28.3) (2022-06-27)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.28.2...v0.28.3)
+
+**Merged pull requests:**
+
+- Add strip\_structures to Insertion Electrode Doc and Builder [\#439](https://github.com/materialsproject/emmet/pull/439) ([acrutt](https://github.com/acrutt))
+
+## [v0.28.2](https://github.com/materialsproject/emmet/tree/v0.28.2) (2022-06-10)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.28.1...v0.28.2)
+
+**Merged pull requests:**
+
+- Add back legacy elasticity model [\#435](https://github.com/materialsproject/emmet/pull/435) ([munrojm](https://github.com/munrojm))
+
+## [v0.28.1](https://github.com/materialsproject/emmet/tree/v0.28.1) (2022-06-08)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.28.0...v0.28.1)
+
+**Merged pull requests:**
+
+- Better handle formula and element issues [\#434](https://github.com/materialsproject/emmet/pull/434) ([munrojm](https://github.com/munrojm))
+
+## [v0.28.0](https://github.com/materialsproject/emmet/tree/v0.28.0) (2022-06-07)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.27.8...v0.28.0)
+
+**Merged pull requests:**
+
+-  Elasticity doc and builder [\#326](https://github.com/materialsproject/emmet/pull/326) ([mjwen](https://github.com/mjwen))
+
+## [v0.27.8](https://github.com/materialsproject/emmet/tree/v0.27.8) (2022-06-07)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.27.7...v0.27.8)
+
+**Merged pull requests:**
+
+- Bump maggma [\#432](https://github.com/materialsproject/emmet/pull/432) ([munrojm](https://github.com/munrojm))
+
+## [v0.27.7](https://github.com/materialsproject/emmet/tree/v0.27.7) (2022-06-06)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.27.6...v0.27.7)
+
+**Merged pull requests:**
+
+- Handle element and formula processing errors [\#431](https://github.com/materialsproject/emmet/pull/431) ([munrojm](https://github.com/munrojm))
+
+## [v0.27.6](https://github.com/materialsproject/emmet/tree/v0.27.6) (2022-05-31)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.27.5...v0.27.6)
+
+**Merged pull requests:**
+
+- Fix facet memory issue with synthesis aggregation pipeline [\#430](https://github.com/materialsproject/emmet/pull/430) ([munrojm](https://github.com/munrojm))
+- Update the description for chemsys\_permutations [\#424](https://github.com/materialsproject/emmet/pull/424) ([acrutt](https://github.com/acrutt))
+
 ## [v0.27.5](https://github.com/materialsproject/emmet/tree/v0.27.5) (2022-05-24)
 
 [Full Changelog](https://github.com/materialsproject/emmet/compare/v0.27.4...v0.27.5)
@@ -282,19 +443,19 @@
 
 ## [v0.21.12](https://github.com/materialsproject/emmet/tree/v0.21.12) (2022-01-19)
 
-[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.21.10...v0.21.12)
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.21.11...v0.21.12)
 
 **Merged pull requests:**
 
 - Revert to `get_distribution` in init [\#349](https://github.com/materialsproject/emmet/pull/349) ([munrojm](https://github.com/munrojm))
 
-## [v0.21.10](https://github.com/materialsproject/emmet/tree/v0.21.10) (2022-01-19)
-
-[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.21.11...v0.21.10)
-
 ## [v0.21.11](https://github.com/materialsproject/emmet/tree/v0.21.11) (2022-01-19)
 
-[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.21.9...v0.21.11)
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.21.10...v0.21.11)
+
+## [v0.21.10](https://github.com/materialsproject/emmet/tree/v0.21.10) (2022-01-19)
+
+[Full Changelog](https://github.com/materialsproject/emmet/compare/v0.21.9...v0.21.10)
 
 **Merged pull requests:**